Check before stripping "exif" off the jp2box
When parsing a jp2 like:

https://tile.loc.gov/storage-services/public/gdcmassbookdig/commandersofarmy00hudn/commandersofarmy00hudn_0009.jp2

We get an error because the EXIF  box doesn't have 'EXIF\0\0' in the beginning - so jp2box._parse_raw_data() skips 6 bytes from the data. When the bytes are not skipped, the EXIF data is properly read.  It causes a warning during parsing in tiff_header(), but causes a problem later if one uses `str(jp2)`, e.g.,

```
 import glymur
 jp2 = glymur.Jp2k("/processing/nlb/image-services-cache/public/gdcmassbookdig/commandersofarmy00hudn/commandersofarmy00hudn_0009.jp2")
/usr/local/lib/python3.8/dist-packages/glymur/jp2box.py:3379: UserWarning: The byte order indication in the TIFF header (b'\x00\x08') is invalid.  It should be either b'II' or b'MM'.
  warnings.warn(str(error))
str(jp2)
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/usr/local/lib/python3.8/dist-packages/glymur/jp2k.py", line 456, in __str__
    metadata.append(str(box))
  File "/usr/local/lib/python3.8/dist-packages/glymur/jp2box.py", line 3439, in __str__
    if tag in self.data:
TypeError: argument of type 'NoneType' is not iterable
```

This change checks that the raw_data starts with exif, when deciding what to pass to tiff_header().